store: Use an async connection pool (SQUASH FROM HERE)
We use an async connnection pool, but only with a synchronous wrapper so
that most uses of connections are not affected by the change yet. Over
time, we will change code to use AsyncPgConn and make it truly async.

Unfortunatley, this change also involves a lot of line noise because of the
replacement of diesel's PgConnection with our own type definition that uses
a AsyncConnectionWrapper

Unfortunately, from this point on, tests will fail because
AsyncConnectionWrapper calls tokio::block_on which happens too often on the
main tokio executor threads. They can only pass once we do not use the
wrapper anymore.